Catch errors emitted during `parinfer-rust--track-undo`
This is meant as a fix for issue #5 where after a users performs a lot
of undos parinfer-rust-mode would stop working. It was stopped working
because when the suer was performing a lot of undo, they would get to
the end of the undo range and get the message "No further undo
information". This message turned out to be a signal from undo that it
had reached an error state. This forced an early exit out of the undo
function and canceling all future function call, including one in the
advice set for after undo was called.

The fix he is to catch the error from undo and ignore it, allowing us
to set our flags after undo is done.