Add option to prompt before buffer change on first load.
Sometimes, when parinfer-rust-mode is first started in a buffer it
needs to modify the indentation to be able to run. This can have
unintended consequences when you're editing in a collaborative
environment and change someone else's specifically chosen
indentation. parinfer-rust-mode now prompts you before it changes the
buffer without your knowing. This option can be customized by setting
the variable `parinfer-rust-check-before-enable` to 't or nil. By
default it's set to 't, so as to not surprise the user with changes to
their files.