fix: prevent buffer jumping around during fast strategy
When using `delete-region` and `insert-buffer-substring` this
sometimes causes the buffer to recenter the text edited area to the
middle or the bottom of the buffer. By tracking the window position
this allows parinfer-rust to make sure the jumping around isn't
minimized. Jumping now only occurs if the user has
`pixel-scroll-precision-mode` enabled and the buffer's start
position is halfway between a line.