fix: when started in a modified state parinfer-rust will not work
When buffer is in a modified state and
`parinfer-rust-check-before-enable` is set to `'defer` then
first-change-hook won't fire while parinfer-rust-mode is enabled.

We can use pre-command-hook to defer running parinfer-rust-mode until
the user interacts with the buffer. This has the issue where, if a
user is just navigating a buffer, parinfer-rust will still need to
run.

As a TODO for a future work we can optimize this by looking for
changes in the buffer and only progressing through check for issue
just before a change is going to be made.

Switch to precommand hook