chore: add flyspell to troublesome modes and minor code cleanup
Flyspell and parinfer-rust-mode don't play well together. Flyspell
moves the cursor around the buffer when spell checking the buffer and,
for some unknown reason, when `parinfer-rust--execute` is executed it
can detect the cursor position as it was being moved around by
flyspell. In safe mode Emacs is smart enough to handle this and put
the cursor in the right place. However, fast mode will just let the
cursor jump around. So, in order to avoid this flyspell-mode should be
disabled if parinfer-rust is to behave properly