fix: prevent buffer jump when safe strategy fails
When `replace-buffer-contents` fails to do a non-destructive replacement
within the allotted time/cost, it basically follows the `fast` strategy,
except unlike the `fast` strategy, it doesn't try to prevent the window
from jumping.

So in addition to doing the window adjustment in the `fast` strategy, we
also do it for the `safe` strategy when it fails and returns `nil`.