fix: bug where `this-command` is nil
This bug was introduce after switching to `track-changes`
library. This is because we've relinquished control for when
parinfer-rust-execute is called. `track-changes` calls it whenever it
feels like it. That means `parinfer-rust-execute` can be called during
a command execution, or after. We still have to prefer `this-command`
first, because `last-command` has out of date information if
`this-command` is not nil.